+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [4.1.1] - Unreleased
 
+- Added: You can now import a game layout/spoiler file in multiworld sessions. 
 - Fixed: The error message when the client is incompatible is now properly displayed.
 - Fixed: Player inventory is now properly sent to the server in multiworld sessions.